Teno-Streamlit/requirements.txt
langchain==0.3.25
langchain_core==0.3.59
langgraph==0.4.3
langsmith==0.3.42
python-dotenv==1.1.0
sentence_transformers==4.1.0
streamlit==1.45.0
torch==2.6.0+cu118
typing_extensions==4.13.2
